Fix nasty bug.
The struct problem used to have a clone of the struct domains' object manager.
That caused that domains pointers to conatant terms where different than
problems' pointers to the same constant. That is a problem in the new
ordered list of children in the tNode (trie), because they are ordered accoring
to the value of their pointer.